requirements/base.txt
Django==3.2.11
django-environ==0.8.1
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-filter==21.1
django-autoslug==1.9.8
python-slugify==6.1.2
django-countries==7.2.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
drf-yasg==1.20.0

Pillow==9.0.0

argon2-cffi==21.3.0

pytz==2022.1

redis==4.1.0
celery==5.2.3
flower==1.0.0
django-celery-email==3.0.0

argh==0.26.2
pyYAML==6.0
watchdog==2.1.6

djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0

django-haystack==3.1.1
Whoosh==2.7.4
drf-haystack==1.8.11